fix: project audit — case-mismatch, deprecated configs, leak-prone defaults
Critical:
- demo/app.py + visualize.py + dataset.create_dummy_manifest: lowercase city
  keys to match real manifests ("warsaw"/"prague"/"budapest"); previously
  capitalized keys silently dropped map markers on real checkpoints.
- evaluate.py + inference.py: defensive case-insensitive class_names sync
  (logs missing-city warning instead of mapping all samples to class 0).
- Delete configs/{config,baseline_config,geoclip_config}.yaml — nested schema
  incompatible with flat TrainConfig; would silently fall back to defaults.

Medium:
- train.py: promote train_frac/val_frac from getattr-hidden fields to
  explicit TrainConfig dataclass fields.
- evaluate.py: drop unused compute_all_metrics import.
- train.py: rename wandb_project "ua-street-geolocation" → match 3-city scope.
- tests/test_dataset.py: align with lowercase dummy cities.

Small:
- pytest.ini: register @pytest.mark.slow marker (kills 2 warnings).
- requirements.txt: rich upper bound <14.0 → <15.0 (14.x is installed).
- .gitignore: ignore RunPod upload artifacts; un-ignore prebuilt manifests.

Verified: 42 passed, 1 skipped.

+++ code/metrics.py
@@ -250,9 +250,9 @@ def _logits_labels_to_preds(
 ) -> tuple[np.ndarray, np.ndarray]:
     """(logits, labels) → (pred_indices, labels) як int64 numpy-масиви."""
     if isinstance(logits, torch.Tensor):
-        logits_np = logits.detach().cpu().numpy()
+        logits_np = logits.detach().cpu().float().numpy()
     else:
-        logits_np = np.asarray(logits)
+        logits_np = np.asarray(logits, dtype=np.float32)
     if isinstance(labels, torch.Tensor):
         labels_np = labels.detach().cpu().numpy().astype(np.int64)
     else:
